+++ packages/vuikit/src/library/modal/components/modal-full.js
@@ -26,7 +26,7 @@ export default {
     Object.keys(this.$slots).forEach(slot => each(this.$slots[slot], node => {
       if (node.fnOptions && node.fnOptions.name === 'VkModalFullClose') {
         assign(node.data, {
-          on: { click: e => this.$emit(TOGGLE, false) }
+          on: assign({ click: e => this.$emit(TOGGLE, false) }, node.data.on || {})
         })
       }
     }))