+++ packages/vuikit/src/table/components/index.js
@@ -1,8 +1,6 @@
 // tables
 export { default as Table } from './table'
-export { default as TableTree } from './table-tree'
 
 // columns
 export { default as TableColumn } from './column'
-export { default as TableColumnTree } from './column-tree'
 export { default as TableColumnSelect } from './column-select'