+++ packages/vuikit/src/table/constants.js
@@ -1,4 +1,0 @@
-export const ROW_LEVEL = '__vkTable_rowLevel'
-export const ROW_CHILDREN_COUNT = '__vkTable_rowChildrenCount'
-export const ROW_CLICK_PREVENTED = '__vkTable_rowClickPrevented'
-export const UPDATE_EXPANDEDROWS = 'update:expandedRows'