+++ packages/vuikit/src/table/index.js
@@ -1,7 +1,3 @@
-import * as constants from './constants'
-
-export { constants }
-
 export * from './elements'
 export * from './components'