+++ packages/vuikit/src/skeleton/__dev__/stories.js
@@ -1,4 +1,0 @@
-import { storiesOf } from '@storybook/vue'
-
-storiesOf('Skeleton', module)
-  .add('Default', () => require('./default.vue').default)