+++ packages/vuikit/src/skeleton/components.js
@@ -1,3 +1,0 @@
-export { ElSkeletonImage as SkeletonImage } from './elements'
-export { ElSkeletonText as SkeletonText } from './elements'
-export { ElSkeletonTitle as SkeletonTitle } from './elements'