+++ packages/vuikit/src/skeleton/elements/index.js
@@ -1,4 +1,0 @@
-export { default as ElSkeleton } from './skeleton'
-export { default as ElSkeletonImage } from './skeleton-image'
-export { default as ElSkeletonText } from './skeleton-text'
-export { default as ElSkeletonTitle } from './skeleton-title'