+++ packages/vuikit/src/skeleton/index.js
@@ -1,2 +1,0 @@
-export * from './elements'
-export * from './components'