+++ packages/vuikit/src/scroll/index.js
@@ -1,3 +1,9 @@
-import directive from './directive'
+import Scroll from './directive'
 
-export default directive
+export { Scroll }
+
+export default {
+  install (Vue, { prefix = 'Vk' } = {}) {
+    Vue.directive(`${prefix}Scroll`, Scroll)
+  }
+}