+++ packages/vuikit/src/list/__dev__/default.vue
@@ -1,0 +1,45 @@
+<template>
+  <div class="uk-padding">
+
+    <h1>List</h1>
+
+    <vk-list>
+      <li>List item 1</li>
+      <li>List item 2</li>
+      <li>List item 3</li>
+    </vk-list>
+
+    <h2>Spaced</h2>
+
+    <vk-list spaced>
+      <li>List item 1</li>
+      <li>List item 2</li>
+      <li>List item 3</li>
+    </vk-list>
+
+    <h2>Divided</h2>
+
+    <vk-list divided>
+      <li>List item 1</li>
+      <li>List item 2</li>
+      <li>List item 3</li>
+    </vk-list>
+
+    <h2>Striped</h2>
+
+    <vk-list striped>
+      <li>List item 1</li>
+      <li>List item 2</li>
+      <li>List item 3</li>
+    </vk-list>
+
+    <h2>Bullet</h2>
+
+    <vk-list bullet>
+      <li>List item 1</li>
+      <li>List item 2</li>
+      <li>List item 3</li>
+    </vk-list>
+
+  </div>
+</template>