+++ packages/vuikit/src/list/__dev__/stories.js
@@ -1,0 +1,4 @@
+import { storiesOf } from '@storybook/vue'
+
+storiesOf('List', module)
+  .add('Default', () => require('./default.vue').default)