+++ packages/vuikit/src/list/components.js
@@ -1,0 +1,1 @@
+export { ElList as List } from './elements'