+++ packages/vuikit/CHANGELOG.md
@@ -14,6 +14,7 @@
    - Added `vk-subnav-item-dropdown-nav`.
    - Removed `vk-nav-dropdown`.
  - Allow `vk-icon-*` to receive the icon as a component.
+ - Add support for individual installs with prefix option.
 
 ## 0.8.5 (May 29, 2018)